requirements.txt
requests==2.26.0
python-dateutils==2.8.2
cerberus==1.3.4
sentry_sdk==1.5.1